// restaraunt-backend/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const RoleType = {
  ADMIN: 'ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN',
  CASHER: 'CASHER',
  WAITER: 'WAITER',
  OWNER: 'OWNER'
} as const

export type RoleType = (typeof RoleType)[keyof typeof RoleType]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  DEBT: 'DEBT'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const WithdrawType = {
  INCOME: 'INCOME',
  OUTCOME: 'OUTCOME'
} as const

export type WithdrawType = (typeof WithdrawType)[keyof typeof WithdrawType]


export const RestaurantType = {
  FAST_FOOD: 'FAST_FOOD',
  CAFE: 'CAFE',
  RESTAURANT: 'RESTAURANT',
  PIZZERIA: 'PIZZERIA',
  SUSHI_BAR: 'SUSHI_BAR',
  OTHER: 'OTHER'
} as const

export type RestaurantType = (typeof RestaurantType)[keyof typeof RestaurantType]